requirements.txt
Pillow==3.0.0
argparse==1.2.1
future==0.16.0
matplotlib==2.0.0
numpy==1.14.3
scipy==0.19.0
tensorflow-gpu==1.4
